relocate: space charge validation functions to validation.py
